libmapper: use size_t
Use the correct type (size_t) for array sizes.

+++ libmapper/internal.h
@@ -1,4 +1,5 @@
 #pragma once
-int sarraylen(char* array[]);
+#include <stddef.h>
+size_t sarraylen(char* array[]);
 void sarrayfree(char* array[]);
 char** sarraydup(char* array[]);
